Improve README and simplify code a bit
This adds an `Internals` section to the README, which gives a high-level
overview of how the generator works.

I also added a few comments to the code and removed some confusing bits
that didn't seem to be used for anything. e.g. `lib.preambule` is gone,
as it was always empty.